feat(pflip): Add pflip command and fix several bugs in rpict
I tested all of these changes against the new "watch the sky" component. There were quite a lot of bugs I found in the rpict command but I believe this fixes them and I was successfully able to get the command to run from a .vf file in my tests.