fix(options): add support for % in radiance options
Several radiance parameters accept values with % as an input. The challenge is that Windows and Unix command line parser work differently with %. In Windows one needs to use %% to make % literal. In Unix it should be passed as %.

To address this issue we will pass %% by default and then find and replace them with % before writing them to Radiance.